remove openai package
moved all code to discord package
